phone_service/requirements.txt
Django>=5.1,<6.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
mysqlclient  # Required for MySQL database connections
httpx  # Used in API Gateway for request forwarding
gunicorn  # For production server
